read/macho: split into modules
